fix: migrate methods still used by tests to non legacy endpoint.
